Исправил : приведение типов(int<->UINT);
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@19 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/apps/mplayerc/VolumeCtrl.cpp
@@ -263,9 +263,9 @@ void CVolumeCtrl::HScroll(UINT nSBCode, UINT nPos)
 {
 //INS:2233 bobdynlan: invalidate
 //-----------------------------------------------------------------------------
-	int nVolMin,nVolMax;
-	GetRange(nVolMin,nVolMax);
-	if (nVolMin <= nSBCode && nSBCode <= nVolMax)
+	int nVolMin, nVolMax;
+	GetRange(nVolMin, nVolMax);
+	if ((UINT)nVolMin <= nSBCode && nSBCode <= (UINT)nVolMax)
 	{
 		CRect r;
 		GetClientRect(&r);
